Customers basic info into separate component
