+++ src/modules/customers/components/detail/sidebar/index.js
@@ -2,5 +2,12 @@ import TaggerSection from './TaggerSection';
 import MessengerSection from './MessengerSection';
 import TwitterSection from './TwitterSection';
 import FacebookSection from './FacebookSection';
+import BasicInfo from './BasicInfo';
 
-export { FacebookSection, TwitterSection, MessengerSection, TaggerSection };
+export {
+  FacebookSection,
+  TwitterSection,
+  MessengerSection,
+  TaggerSection,
+  BasicInfo
+};
